# Using generators in sabre/vobject where they make sense