Fixes #282