@evert 2015-12-12
Benchmarks

Before

```
__global__  |||||||||||||||||||||||||||||||||||||||||||||||||||| 57830ms, 100.0%
parse                                                              257ms,   0.4%
fb          |||||||||||||||||||||||||||||||||||||||||||||||||||| 57573ms,  99.6%

peak memory usage: 14M
current memory usage: 13M
```

After

```
__global__  |||||||||||||||||||||||||||||||||||||||||||||||||||| 57561ms, 100.0%
parse                                                              256ms,   0.4%
fb          |||||||||||||||||||||||||||||||||||||||||||||||||||| 57305ms,  99.6%

peak memory usage: 15M
current memory usage: 13M
```

Unfortunately this implies there's nearly no performance benefit, and slightly more memory usage. Because of this, I'll discard the patch.